moving everything individually
