+++ apps/client/components/userProfile/prize-tab.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function PrizeTab({params}: {params: {id: string}}) {
+  return (
+    <div>
+      
+    </div>
+  )
+}
